Time: 16 ms (55.47%), Space: 10.9 MB (59.74%) - LeetHub

+++ 462-minimum-moves-to-equal-array-elements-ii/462-minimum-moves-to-equal-array-elements-ii.cpp
@@ -4,10 +4,10 @@ public:
         
         sort(a.begin(), a.end());
         int sum = a[a.size()/2];
-        long long ans = 0;
+        int ans = 0;
         for(int i = 0; i< a.size(); i++)
             ans += abs(sum-a[i]);
         
-        return (int)ans;
+        return ans;
     }
 };
